mentor/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
langchain-core